Fix architecture-vision Job numbering and move SandboxClient off the k8s adapter
Two documentation/structure defects from the #541 crawl, fenced off there:

1. architecture-vision.md numbered its jobs one way in the prose (### 1 Harness
   ... ### 6 Channel) and another in its own mermaid diagram (Job 1 = obs, Job 5
   = harness), so a reader following 'Job 3' landed on the wrong box. Relabeled
   the diagram to match the canonical ordered prose sections.
2. The SandboxClient Protocol -- the substrate seam ARCHITECTURE.md calls 'the
   load-bearing design property of the whole system' -- was declared inside the
   concrete k8s adapter (sandbox/k8s.py), reading as 'the port is a Kubernetes
   thing'. Moved it (with OperatingMode, to avoid a types<->k8s circular import)
   to the substrate-neutral sandbox/types.py; k8s.py/docker.py/substrate.py/the
   package __init__ import it from there. Pure move, no behavior change.

Ref CUR-1625
Closes #543
