+++ apps/worker/src/agentos_worker/sandbox/docker.py
@@ -56,9 +56,8 @@ from ..bundle_store import BundleReader, extract_bundle
 from .k8s import (
     MANAGED_BY_LABEL,
     MANAGED_BY_VALUE,
-    OperatingMode,
 )
-from .types import ClaimView, SandboxError, SandboxView
+from .types import ClaimView, OperatingMode, SandboxError, SandboxView
 
 logger = logging.getLogger(__name__)
